Bringing in toastr. Adding some semblance of front-end security. Adding real-time issue preview.

+++ app/templates/TemplateCtrl.js
@@ -1,4 +1,4 @@
-angular.module('it').controller('TemplateCtrl', function($scope, $location, util, TemplateService, template, mode) {
+angular.module('it').controller('TemplateCtrl', function($scope, $location, util, TemplateService, GitHubService, template, mode, $timeout, toastr) {
   $scope.template = template;
   $scope.mode = mode;
 
@@ -49,9 +49,73 @@ angular.module('it').controller('TemplateCtrl', function($scope, $location, util
     };
   })(false);
 
+  /*
+   * Collaborator check logic
+   */
+  $scope.isCollaborator = $scope.mode === 'new';
+
+  function checkIfUserIsCollaborator() {
+    if ($scope.template.owner && $scope.template.repo && $scope.user && $scope.user.login) {
+      GitHubService.checkUserIsCollaborator($scope.template.owner, $scope.template.repo, $scope.user.login).then(function(isCollaborator) {
+        $scope.isCollaborator = isCollaborator;
+        if (!isCollaborator) {
+          showCollabWarning();
+        }
+      });
+    } else {
+      $scope.isCollaborator = false;
+    }
+  }
+  var timeout;
+
+  function checkTimeout() {
+    $timeout.cancel(timeout);
+    return $timeout(function() {
+      checkIfUserIsCollaborator();
+      if (mode !== 'edit') {
+        checkIfOverwrittingTemplate();
+      }
+    }, 500, true);
+  }
+
+  $scope.$watch('template.owner + template.repo + user + template.name', function() {
+    timeout = checkTimeout();
+  });
+
+  function showCollabWarning() {
+    toastr.warning('You are not a collaborator on ' + util.simpleCompile('{{owner}}/{{repo}}.' +
+      ' You wont be able to save this template.', $scope.template), 'Warning...');
+  }
+
+
+  /*
+   * Template overwrite logic
+   */
+  function checkIfOverwrittingTemplate() {
+    if ($scope.template.owner && $scope.template.repo && $scope.template.name) {
+      TemplateService.getTemplate($scope.template).once('value', function(template) {
+        if (template.val()) {
+          showOverwriteWarning();
+        }
+      });
+    }
+  }
+
+  function showOverwriteWarning() {
+    toastr.warning('There is already a template at ' + util.simpleCompile('{{owner}}/{{repo}}/{{name}}.' +
+      ' Saving this template will overwrite that template.', $scope.template), 'Warning...');
+  }
+
+  /*
+   * Form controls
+   */
   $scope.deleteTemplate = function() {
-    TemplateService.deleteTemplate($scope.template);
-    $location.path('/');
+    if ($scope.isCollaborator) {
+      TemplateService.deleteTemplate($scope.template);
+      $location.path('/');
+    } else {
+      showCollabWarning();
+    }
   };
 
   $scope.copyTemplate = function() {
@@ -68,9 +132,13 @@ angular.module('it').controller('TemplateCtrl', function($scope, $location, util
   };
 
   $scope.submitTemplate = function() {
-    $scope.template.createdBy = $scope.user.login;
-    TemplateService.saveTemplate(JSON.parse(angular.toJson($scope.template)));
-    $scope.templateUrl = util.simpleCompile('#/:owner/:repo/:name', $scope.template);
+    if ($scope.isCollaborator) {
+      $scope.template.createdBy = $scope.user.login;
+      TemplateService.saveTemplate(JSON.parse(angular.toJson($scope.template)));
+      $scope.templateUrl = util.simpleCompile('#/{{owner}}/{{repo}}/{{name}}', $scope.template);
+    } else {
+      showCollabWarning();
+    }
   };
 
 });
